enable vaapi support
Unfortunately, having vaapi means that we can no longer make an
absolutely static binary -- because libva does not support being linked
statically.

The build system of ffmpeg (or just autotools in general) does not
really handle this "mostly static" scenario very well. We have to use a
lot of dirty hacks. In particular, we have to pass `-Wl,-Bstatic` in
extra ld flags instead of `-static` in order to make most libraries
statically linked, while retaining the ability to use `-Wl,-Bdynamic` to
link part of the libraries dynamically. We can no longer rely on the
`--static` parameter of `pkg-config` because some libraries may refer to
things like `libm` in their static `pkg-config` while we can no longer
link these libraries statically (because the existence of a dynamic
`libva`, linking these libraries statically will result in two different
versions being loaded into memory at the same time).

The `configure` script also needs to be patched to ignore `-lva` which
is required by the `vaapi` option -- because we need to place it after
`-Wl,-Bdynamic` in order to make it link dynamically. The check of API
correctness is also bypassed on this library because it causes problems
and who cares.

Currently dynamically-linked libraries:

```
[peter@peter-pc ffmpeg-static]$ ldd out/ffmpeg
        linux-vdso.so.1 (0x00007ffde3f86000)
        libdl.so.2 => /usr/lib/libdl.so.2 (0x00007f4ec0122000)
        libpthread.so.0 => /usr/lib/libpthread.so.0 (0x00007f4ec0100000)
        libm.so.6 => /usr/lib/libm.so.6 (0x00007f4ebffba000)
        libmvec.so.1 => /usr/lib/libmvec.so.1 (0x00007f4ebff8e000)
        libc.so.6 => /usr/lib/libc.so.6 (0x00007f4ebfdc8000)
        /lib64/ld-linux-x86-64.so.2 => /usr/lib64/ld-linux-x86-64.so.2 (0x00007f4ec33f8000)
        libva.so.2 => /usr/lib/libva.so.2 (0x00007f4ebfda3000)
```

I don't expect these libraries to be updated very often anyway so should
be fine under most circumstances...